Add ByesCommand to application
In order to search for or get a list of bye weeks, this commit adds the
ByesCommand. The byes command takes one optional argument of the week
number and will return all the teams that are on a bye that week.  If no
week was passed to the command it lists all teams with bye weeks.