Add ScheduleCommand
In order to get a schedule for a team or a week this commit adds a
`schedule` command to the `ffnerd` application. This commit also adds a
default schedule of current week to be returned.  This commit fixes a
test for `empty()` function instead of using the `!=` for a variable in
all commands.  This commit add a `Game` model to store specific game
schedule data.